feat: add individual month / week / day modules now you can just import what you need to help reduce bundle size Closes #361
