+++ src/common/calendar-event-title-formatter.provider.ts
