+++ src/common/calendar-resize-helper.provider.ts
