+++ src/common/click.directive.ts
