+++ src/day/calendar-day-view-hour-segment.component.ts
