+++ src/index.ts
@@ -1,20 +1,1 @@
-export * from './providers/calendar-event-title-formatter.provider';
-export * from './providers/calendar-moment-date-formatter.provider';
-export * from './providers/calendar-native-date-formatter.provider';
-export * from './providers/calendar-date-formatter.provider';
-export * from './providers/calendar-utils.provider';
-export * from './interfaces/calendar-date-formatter.interface';
-export * from './interfaces/calendar-event-times-changed-event.interface';
-export * from './modules/calendar';
-export * from './components/day/calendar-day-view.component';
-export * from './components/month/calendar-month-view.component';
-export * from './components/week/calendar-week-view.component';
-export {
-  CalendarEvent,
-  EventAction as CalendarEventAction,
-  MonthViewDay as CalendarMonthViewDay,
-  WeekViewEvent as CalendarWeekViewEvent,
-  WeekViewEventRow as CalendarWeekViewEventRow,
-  GetWeekViewArgs as CalendarGetWeekViewArgs,
-  DAYS_OF_WEEK
-} from 'calendar-utils';
+export * from './calendar.module';
