+++ src/month/calendar-month-view.component.ts
@@ -28,9 +28,9 @@ import getMonth from 'date-fns/get_month';
 import getYear from 'date-fns/get_year';
 import differenceInSeconds from 'date-fns/difference_in_seconds';
 import addSeconds from 'date-fns/add_seconds';
-import { CalendarEventTimesChangedEvent } from '../../interfaces/calendar-event-times-changed-event.interface';
-import { CalendarUtils } from '../../providers/calendar-utils.provider';
-import { validateEvents } from '../../providers/util';
+import { CalendarEventTimesChangedEvent } from '../common/calendar-event-times-changed-event.interface';
+import { CalendarUtils } from '../common/calendar-utils.provider';
+import { validateEvents } from '../common/util';
 
 /**
  * Shows all events on a given month. Example usage:
