+++ src/month/calendar-month.module.ts
@@ -1,0 +1,28 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { DragAndDropModule } from 'angular-draggable-droppable';
+import { CalendarMonthViewComponent } from './calendar-month-view.component';
+import { CalendarMonthViewHeaderComponent } from './calendar-month-view-header.component';
+import { CalendarMonthCellComponent } from './calendar-month-cell.component';
+import { CalendarOpenDayEventsComponent } from './calendar-open-day-events.component';
+import { CalendarCommonModule } from '../common/calendar-common.module';
+
+export { CalendarMonthViewComponent } from './calendar-month-view.component';
+export { MonthViewDay as CalendarMonthViewDay } from 'calendar-utils';
+
+@NgModule({
+  imports: [CommonModule, DragAndDropModule, CalendarCommonModule],
+  declarations: [
+    CalendarMonthViewComponent,
+    CalendarMonthCellComponent,
+    CalendarOpenDayEventsComponent,
+    CalendarMonthViewHeaderComponent
+  ],
+  exports: [
+    CalendarMonthViewComponent,
+    CalendarMonthCellComponent,
+    CalendarOpenDayEventsComponent,
+    CalendarMonthViewHeaderComponent
+  ]
+})
+export class CalendarMonthModule {}
