+++ src/week/calendar-week-view-event.component.ts
