+++ src/week/calendar-week-view.component.ts
@@ -21,11 +21,11 @@ import {
 } from 'calendar-utils';
 import { ResizeEvent } from 'angular-resizable-element';
 import addDays from 'date-fns/add_days';
-import { CalendarDragHelper } from '../../providers/calendar-drag-helper.provider';
-import { CalendarResizeHelper } from '../../providers/calendar-resize-helper.provider';
-import { CalendarEventTimesChangedEvent } from '../../interfaces/calendar-event-times-changed-event.interface';
-import { CalendarUtils } from '../../providers/calendar-utils.provider';
-import { validateEvents } from '../../providers/util';
+import { CalendarDragHelper } from '../common/calendar-drag-helper.provider';
+import { CalendarResizeHelper } from '../common/calendar-resize-helper.provider';
+import { CalendarEventTimesChangedEvent } from '../common/calendar-event-times-changed-event.interface';
+import { CalendarUtils } from '../common/calendar-utils.provider';
+import { validateEvents } from '../common/util';
 
 export interface WeekViewEventResize {
   originalOffset: number;
