+++ src/week/index.ts
@@ -1,0 +1,1 @@
+export * from './calendar-week.module';
